chore(sdk-rev2.0): v2.3.0 FW 동기화 — 헤더/Core/Compatible/LWIP/libXM_Lib.a 갱신
XM-Dev(Develop_Rev2.0) 최신 상태로 SDK Rev2.0 전체 동기화 — 릴리즈 SDK가
개발 펌웨어와 완전히 동일하게 빌드·동작하도록 정합.

- XM_FW 헤더 39종 갱신 + 7종 추가 + libXM_Lib.a 재빌드
  · 디버그정보 개발경로 스크럽(-ffile-prefix-map, 1007→0)
  · DWARF zlib 압축(objcopy --compress-debug-sections) — GitHub 100MB blob
    한도 대응(107.9→47.3MB). nm 심볼 9,029 동일·비-debug 섹션 23,914 동일·
    링크 산출 .bin SHA-256 동일(동작 완전 동일)
- Core(main.c/freertos.c/stm32h7xx_it.c/hal_msp.c/FreeRTOSConfig.h/system_stm32h7xx.c)
  + linker 동기화 — 그동안 미반영되던 NVIC/우선순위/fault 핸들러 정합
- Compatible/LWIP 컴파일 소스 정합, 예제 50종 + Ex.40/41/42 SDK 동봉
- tools/codegen/data_map 동기화(yaml v2.5→v2.6 + 생성기 + .ts v2.4→v2.6) —
  구버전 yaml로 빌드 시 커밋된 v2.6 헤더가 v2.5로 역재생성되던 결함 수정
  (User_Custom 28B 슬롯 소실). 재생성 안정성 검증(출력=커밋본, 타임스탬프 제외)
- stale 헤더 4종 제거(CoE·node_id·task_state_machine 이동)
- version.h → 2.3.0.0
- 검증: 릴리즈 SDK 클린빌드 0 error (arm-none-eabi-gcc 13.3.1)

+++ XM10_SDK/Rev2.0/Extension_Module/XM_FW/AGR_MW/DOP/Core/agr_pdo_engine.h
@@ -143,7 +143,7 @@ int32_t AGR_PDO_ApplyMapFromSDO(AGR_PDO_MapTable_t* map,
 int32_t AGR_PDO_Encode(const AGR_PDO_MapTable_t* map,
                    const AGR_OD_Table_t* od,
                    uint8_t* out_buf,
-                   uint8_t buf_size);
+                   uint16_t buf_size);
 
 /**
  * @brief PDO 페이로드 디코딩 (바이트 버퍼 → OD Entry)
